Изменен файл DuckActionsClient.java Добавлен файл DuckPropertiesClient.java
